Formalizes TickFunc. No longer runs a tick in a separate goro.
It is known that starting a new goro and executing a function in it
has variable and non-trivial time costs. For this application, after
a lot of testing, the best-case gains are not worth the average-case
losses. The largest negative side-effect is that if the TickFuncs
take longer than an interval, the interval will be missed. This can
be mitigated with a deadline, but that's going to kill elegance.